+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - `Issue`
     - `#signature` -- Store `Regexp` source instead of converting it to String.
 - Checks -- Added proofs to as many issues as possible.
+    - `xss`
+        - When the case involves payloads landing in `textarea`s, break out of
+            them to prevent possible FPs.
 - Plugins
     - `autologin` -- Updated to fail gracefully in cases of an invisible form DOM elements.
 - Reporters